requirements.txt
pandas
python-docx
tqdm
